refactor: moved the graph fns to a new file
