+++ champ/node/consensus.rs
@@ -1,1 +1,2 @@
+pub mod graphs;
 pub mod voting_power;
