+++ champ/node/consensus/voting_power.rs
@@ -1,13 +1,9 @@
 use anyhow::{anyhow, Result};
-use pog_proto::api::Block;
 
+use crate::consensus::graphs;
 use crate::state::ChampStateMutex;
-use std::f64::consts::E;
 
-// so we can normalize the curves
-const NORMALIZED_MAX_POW: f64 = 10.0;
-const TX_CURVE_MAX: i32 = 15;
-const PLATEAU_SIZE: f64 = 350.0;
+// To balance each graph
 const HEIGHT_WEIGHT: f64 = 1.0;
 const BALANCE_WEIGHT: f64 = 1.0;
 const AGE_WEIGHT: f64 = 1.0;
@@ -19,8 +15,6 @@ const MAX_NETWORK_POWER: f64 = 0.3;
 const LOOKBACK_RANGE: u64 = 60 * 60 * 24 * 30;
 // 2 Months in Seconds
 const MAX_LOOKBACK_RANGE: u64 = 60 * 60 * 24 * 30 * 2;
-
-const WEEK_IN_SECONDS: f64 = 60.0 * 60.0 * 24.0 * 7.0;
 
 /// Returns actual voting power of an account.
 /// Actual voting power is without the delegated power.
@@ -42,9 +36,9 @@ pub async fn get_actual_power(state: &ChampStateMutex, account_id: String) -> Re
     // First Block from an account
     let first_block = db.get_block_by_height(&account_id, &0).await?;
 
-    let bresult = balance_graph(data.balance);
-    let hresult = tx_graph(data.height, block, old_block_result);
-    let aresult = age_graph(block.timestamp - first_block.timestamp);
+    let bresult = graphs::balance_graph(data.balance);
+    let hresult = graphs::tx_graph(data.height, block, old_block_result);
+    let aresult = graphs::age_graph(block.timestamp - first_block.timestamp);
 
     // TODO: Green Adresses?
 
@@ -59,7 +53,7 @@ pub async fn get_actual_power(state: &ChampStateMutex, account_id: String) -> Re
 #[allow(dead_code)]
 /// Gets the sum of the power of each delegate of an account
 async fn get_delegated_power(state: &ChampStateMutex, account_id: String) -> Result<u32> {
-    // TODO: Cache this 
+    // TODO: Cache this
     let mut power = 0;
     let db = &state.lock().await.db;
 
@@ -79,56 +73,15 @@ async fn get_delegated_power(state: &ChampStateMutex, account_id: String) -> Res
 pub async fn get_active_power(state: &ChampStateMutex, account_id: String) -> Result<u32> {
     let actual_power = get_actual_power(state, account_id.clone()).await?;
     let delegate_power = get_delegated_power(state, account_id.to_owned()).await?;
-    let total_network_power = max_voting_power();
+    let total_network_power = get_max_voting_power();
     let total_power = actual_power + delegate_power;
     if total_power > total_network_power {
         return Ok(total_network_power);
     }
     Ok(total_power)
 }
-
-fn balance_graph(balance: u64) -> f64 {
-    // base function that is an Asymptote with Balance
-    // where x is a balance type (tbd)
-    // This tends towards 10, as the balance increases
-    NORMALIZED_MAX_POW / (1.0 + E.powf((balance as f64 / -5.0) + NORMALIZED_MAX_POW))
-}
-
-fn tx_graph(block_height: u64, new_block: &Block, old_block: Option<&Block>) -> f64 {
-    let old_block_time = match old_block {
-        Some(b) => b.timestamp,
-        None => new_block.timestamp,
-    };
-    // to get the time between the first and most recent block
-    // we need the minimum to not give too high power from the start
-    let time = if new_block.timestamp - old_block_time < WEEK_IN_SECONDS as u64 {
-        WEEK_IN_SECONDS
-    } else {
-        (new_block.timestamp - old_block_time) as f64
-    };
-
-    // x is the nr of tx based on the account life in weeks
-    // https://www.geogebra.org/calculator/ymkv5ew6
-    let tx_per_week = (time / block_height as f64) / WEEK_IN_SECONDS;
-    // this is between 0 and 1 where plateau starts at 0.5
-    let graph_result = 1.0 / (tx_per_week / (PLATEAU_SIZE / 2.0) - 1.0).powi(2 * TX_CURVE_MAX) + 1.0;
-    // to normalize tx graph and balanc graph
-    graph_result * 10.0
-}
-
-fn age_graph(account_age: u64) -> f64 {
-    // x is the account age in weeks
-    // starts with negative power but increases at around 1 month
-    // slowly increases steadily
-    // x + 1 to avoid log0
-    // 0.1x + 3 to allow the graph to go through 31 (month ish)
-    // - 4 to shift the start
-    let account_age_weeks = (account_age as f64 / WEEK_IN_SECONDS).floor();
-    (account_age_weeks + 1.0).log10() + (0.1 * account_age_weeks + 3.0).sqrt() - 4.0
-}
-
 /// Gets the max voting power in the system and sets a limit of a percentage
-fn max_voting_power() -> u32 {
+fn get_max_voting_power() -> u32 {
     //TODO: Get all voting power in the system
     (100_000_000_f64 * MAX_NETWORK_POWER) as u32
 }
